backend/requirements-django.txt
Django>=5.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.1
django-cors-headers>=4.3.1
django-filter>=23.5
drf-simplejwt>=5.3.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.2.0
django-environ>=0.11.2
python-dateutil>=2.8.2
